+++ apps/web/src/actions/campaign/delete-campaign.ts
@@ -1,23 +1,35 @@
 "use server"
 
 import { routes } from "@repo/routes"
+import { SESSION_COOKIE } from "@repo/types/auth"
 
-import { sessionMiddleware } from "~/actions/utils"
-import { getBackendUrl } from "~/lib/url"
+import { actionError, actionSuccess, type ActionResult } from "~/lib/actions"
+import { getCookie } from "~/lib/cookies"
+import { getBackendPrivateUrl } from "~/lib/url"
 
-const deleteCampaign = async (organizationId: string, campaignId: string): Promise<void> => {
-	const sessionToken = await sessionMiddleware()
+const deleteCampaign = async (organizationId: string, campaignId: string): Promise<ActionResult> => {
+	const sessionToken = await getCookie(SESSION_COOKIE)
+	if (!sessionToken) throw new Error("Unauthorized")
 
-	const backendUrl = getBackendUrl()
-	const res = await fetch(`${backendUrl}${routes.backend.DELETE_CAMPAIGN(organizationId, campaignId)}`, {
-		method: "DELETE",
-		headers: {
-			"Authorization": `Bearer ${sessionToken}`
+	const backendUrl = getBackendPrivateUrl()
+
+	try {
+		const res = await fetch(`${backendUrl}${routes.backend.DELETE_CAMPAIGN(organizationId, campaignId)}`, {
+			method: "DELETE",
+			headers: {
+				"Authorization": `Bearer ${sessionToken}`
+			}
+		})
+		if (!res.ok) {
+			const errData = await res.json()
+			return actionError(errData.message)
 		}
-	})
-	if (!res.ok) {
-		const errData = await res.json()
-		throw new Error(errData.message)
+
+		return actionSuccess()
+	} catch (err: unknown) {
+		if (err instanceof Error) console.error(err.message, err.stack)
+
+		return actionError()
 	}
 }
