+++ apps/web/src/actions/contact/delete-many-contacts.ts
@@ -1,26 +1,38 @@
 "use server"
 
 import { routes } from "@repo/routes"
+import { SESSION_COOKIE } from "@repo/types/auth"
 import { type DeleteManyContactsDto } from "@repo/types/contact"
 
-import { sessionMiddleware } from "~/actions/utils"
-import { getBackendUrl } from "~/lib/url"
+import { actionError, type ActionResult, actionSuccess } from "~/lib/actions"
+import { getCookie } from "~/lib/cookies"
+import { getBackendPrivateUrl } from "~/lib/url"
 
-const deleteManyContacts = async (organizationId: string, dto: DeleteManyContactsDto): Promise<void> => {
-	const sessionToken = await sessionMiddleware()
+const deleteManyContacts = async (organizationId: string, dto: DeleteManyContactsDto): Promise<ActionResult> => {
+	const sessionToken = await getCookie(SESSION_COOKIE)
+	if (!sessionToken) throw new Error("Unauthorized")
 
-	const backendUrl = getBackendUrl()
-	const res = await fetch(`${backendUrl}${routes.backend.DELETE_MANY_CONTACTS(organizationId)}`, {
-		method: "DELETE",
-		body: JSON.stringify(dto),
-		headers: {
-			"Authorization": `Bearer ${sessionToken}`,
-			"Content-Type": "application/json"
+	const backendUrl = getBackendPrivateUrl()
+
+	try {
+		const res = await fetch(`${backendUrl}${routes.backend.DELETE_MANY_CONTACTS(organizationId)}`, {
+			method: "DELETE",
+			body: JSON.stringify(dto),
+			headers: {
+				"Authorization": `Bearer ${sessionToken}`,
+				"Content-Type": "application/json"
+			}
+		})
+		if (!res.ok) {
+			const errData = await res.json()
+			return actionError(errData.message)
 		}
-	})
-	if (!res.ok) {
-		const errData = await res.json()
-		throw new Error(errData.message)
+
+		return actionSuccess()
+	} catch (err: unknown) {
+		if (err instanceof Error) console.error(err.message, err.stack)
+
+		return actionError()
 	}
 }
